chore: add scenes for benchmark
